Various changes to permit compiling OpenMM stuff on Linux Added INCLUDE(Dart) to OpenMM CMakeLists.txt

+++ openmmapi/include/System.h
@@ -134,7 +134,7 @@ public:
     }
 private:
     class ConstraintInfo;
-    std::vector<int> masses;
+    std::vector<double> masses;
     std::vector<ConstraintInfo> constraints;
     std::vector<Force*> forces;
 };
